Regroup recyclerview tests
