+++ app/src/test/java/com/example/repertoire/SongRecyclerViewTest.kt
@@ -2,14 +2,15 @@ package com.example.repertoire
 
 import android.os.Looper.getMainLooper
 import androidx.recyclerview.widget.RecyclerView
-import org.junit.Test
-import org.junit.runner.RunWith
 import androidx.test.core.app.launchActivity
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import org.junit.After
 import org.junit.Assert.*
 import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.junit.runners.JUnit4
 import org.robolectric.Shadows.shadowOf
 import org.robolectric.annotation.Config
 
@@ -86,5 +87,61 @@ class SongRecyclerViewTest {
         
         val context = InstrumentationRegistry.getInstrumentation().targetContext
         AppDatabase.getInstanceAllowingMainThreadQueriesForTests(context).close()
+    }
+}
+
+
+@RunWith(JUnit4::class)
+class SongItemCallbackTestTest {
+    private val callback = SongItemCallback()
+    private val song = Song(
+        uri = "content://arbitrary/uri",
+        name = "Pearl Jam - Black"
+    )
+
+    @Test
+    fun identicalSongsAreTheSameItems() {
+        assertTrue(callback.areItemsTheSame(song,song))
+    }
+
+    @Test
+    fun songsWithSameUriAreTheSameItems() {
+        val song2 = Song(
+            uri = song.uri,
+            name = "anotherName"
+        )
+        assertTrue(callback.areItemsTheSame(song,song2))
+    }
+
+    @Test
+    fun songsWithDifferentUriAreNotTheSameItems() {
+        val song2 = Song(
+            uri = "content://anotherArbitrary/uri",
+            name = song.name
+        )
+        assertFalse(callback.areItemsTheSame(song,song2))
+    }
+
+    @Test
+    fun identicalSongsHaveTheSameContents() {
+        assertTrue(callback.areContentsTheSame(song,song))
+    }
+
+    @Test
+    fun songsWithDifferentNameHaveNotTheSameContents() {
+        val song2 = Song(
+            uri = song.uri,
+            name = "anotherName"
+        )
+        assertFalse(callback.areContentsTheSame(song,song2))
+    }
+
+    @Test
+    fun songsWithDifferentUriHaveNotTheSameContent() {
+        val song2 = Song(
+            uri = "content://anotherArbitrary/uri",
+            name = song.name
+        )
+        assertFalse(callback.areContentsTheSame(song,song2))
     }
 }
